refactor: redesigned site and added content
Added more information to all sections of the documentation site,
especially the `skills` and `contact` sections. Updated the static
images to be more colorful and vibrate, edited the homepage to look a
bit more lively as well and removed the `users` section. Need to
continue adding content as needed and adjusting the styling to the
team's desire.